+++ packages/resqlite_js/lib/src/resqlite_js.dart
@@ -2,7 +2,7 @@ import 'dart:ffi';
 
 import 'package:resqlite/resqlite.dart';
 
-@Native<Int Function(Pointer<Void>, Pointer<Void>, Pointer<Void>)>(
+@Native<ResqliteExtensionInitNative>(
   assetId: 'package:resqlite_js/src/native/sqlite_js_extension.dart',
   symbol: 'sqlite3_js_init',
 )
@@ -15,9 +15,7 @@ external int sqlite3JsInit(
 /// Loads SQLite JS on every connection opened by resqlite.
 ResqliteExtension sqliteJsExtension() {
   return ResqliteExtension(
-    Native.addressOf<NativeFunction<ResqliteExtensionInitNative>>(
-      sqlite3JsInit,
-    ),
+    Native.addressOf<ResqliteExtensionEntrypoint>(sqlite3JsInit),
     name: 'sqlite_js',
   );
 }